requirements.txt
dash==2.14.2
plotly==5.18.0
pandas==2.1.4
gunicorn==21.2.0
psycopg2-binary==2.9.9